phase-2-rag/01-basic-rag/requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-text-splitters>=0.2.0
pypdf>=4.0.0
python-dotenv>=1.0.0
litellm>=1.40.0
rich>=13.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
